html/semantics/scripting-1/the-script-element/json-module-service-worker/json-module-service-worker-test.https.html

<!doctype html>

<title>import-json-module-service-worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
    promise_test(function (test) {
        return navigator.serviceWorker.register('./serviceworker.js', { type: 'module' })
            .then(function (reg) {
                //registration successful
                assert_true(true, "Javascript importing JSON Module loaded successfully within the context of a service worker")
            }, function (error) {
                // registration failed, should never reach here
                assert_unreached("Javascript importing JSON Module failed to load within the context of a service worker")
            });
    }, "Javascript importing JSON Module should load within the context of a service worker");
</script>

<script>
    promise_test(function (test) {
        return navigator.serviceWorker.register('./module.json', { type: 'module' })
            .then(function (reg) {
                //registration successful
                assert_true(true, "JSON Module loaded successfully within the context of a service worker")
            }, function (error) {
                // registration failed, should never reach here
                assert_unreached("JSON Module failed to load within the context of a service worker")
            });
    }, "JSON Modules should load within the context of a service worker");
</script>
